Create README.md
This is a non-market Android application / apk for controlling the settings on a custom programmable LED array designed to test lighting preferences of insects. Fabrication files for the LED array are in the repository https://github.com/danielje/Smart-Trap-Lighting-Hardware, with the device firmware compatible with this app available in the repository https://github.com/danielje/Smart-Trap-Lighting-Firmware. A manual with instructions how to use the app to program the device is included as a pdf in this repository.